Update unit tests for `IPv4Addr`.
Prefix the members of the `tests` struct with `z%02d_` in order to
allow for both easy alpha-sorting of struct members, but also to put
struct members in a coherent order that allows for tests to be more
easily scanned visually.  Similarly, order tests by their lexical
sort order now.