Fix RFC4192 addresses passed to NewIPv6Addr()
IPv4-mapped addresses in RFC4291 will now parse as IPv6 addresses.  Previously they
would be parsed successfully as both an IPv6 and IPv4 address, and therefore would be
assumed to be an IPv4 address and would return an error as an IPv6 address.  Now we test
to see if there is a period in the input string and abort early if there is.