Change IfByMaskSize()'s function signature to accept a string selector like the rest of the filter functions.
I did this to allow the ability to specify ranges for network sizes.
For now it's an exact match, but the express intention of this change is
to be able to specify things like `>=28` where it will match any network
whose netmask is a 28 or larger (/32 having a smaller network size, but
a larger netmask).