Actually add Windows support this time (IPv4 only at this time).
The first pass at adding Windows support was a few bricks shy of a load.
In the first pass we correctly identified the IP address of the
interface serving the default route, but incorrectly assumed the
Interface name was the IP address listed in the output from `netstat
-rn`.  In order to actually do the right thing, we need to use that IP
address along with parsing `ipconfig` in order to find the interface
name.

Important to note: only IPv4 is supported at this time.  If someone has
a host connected via IPv6, please send the output to `ipconfig`,
`netstat -rn`, and the version of windows as issues on
github.com/hashicorp/go-sockaddr in order to see IPv6 support added for
Windows.