Add the `math` operation to perform calculations on IPs and networks.
`math` takes two arguments: an operation and a string value. The two initial
operations are `address` and `network`.

The `address` operation's value is a decimal encoded string with a mandatory
sign.  The value is added or subtracted from the IP address of the IfAddr.
Values will overflow or underflow, and if necessary, wrap at the underlying
fundamental type.

The `network` operation's value is a decimal encoded string with a mandatory
sign.  If the value is positive it is added to the network address of the
IfAddr.  If the value overflows the network size, the value is wrapped at the
network boundary (i.e. 257 will wrap in a /24 to be the same as +1).  If the
value is negative it is subtracted from the broadcast address of the IfAddr.
If the value underflows the network size, the value will wrap.

The `network` operation is intended to address the concern raised in #7 but
in a more generic way.