Fix GetPrivateInterfaces() and GetPublicInterfaces() for dual-homed hosts.
When a host has two interfaces, one with a public IP address and another
with a private IP address, it should be possible to select the private IP
address regardless of whether or not the default route is attached to the
interface.  This is a subtle, but likely harmless change.

The following commit will introduce a `sort` option for `default` to address
any remaining concern regarding the sort order of multi-homed hosts with or
without a default route on a given interface.