Teach sort how to sort IfAddrs with the default route.
In multi-homed hosts, the default route may exist on an interface with
the private IP address or on a host with the public IP address.  To
handle this uncertainty, add the appropriate sort helper function to
enable sorting of IfAddrs such that the IfAddr who has a default route
will sort first (or last).  This allows for correct ordering of
GetPrivateIP and GetPublicIP.