Add a marshaling/unmarshaling convenience type. (#14)
It's often nice to marshal structs that contain SockAddrs; it's less
nice to have to write your own marshaling logic for each of these. This
adds a struct that merely contains a SockAddr and JSON marshal/unmarshal
functions on the struct so that this struct can be used directly in
other types.